@@ -1,6 +1,6 @@
 [tool.black]
 line-length = 100
-target-version = ['py38']
+target-version = ['py39']
 exclude = 'generated'
 
 [tool.isort]
