@@ -15,7 +15,7 @@ from custom_components.hacs.const import HACS_ACTION_GITHUB_API_HEADERS
 from custom_components.hacs.enums import HacsGitHubRepo
 from custom_components.hacs.exceptions import HacsException
 from custom_components.hacs.utils.decode import decode_content
-from custom_components.hacs.utils.logger import get_hacs_logger
+from custom_components.hacs.utils.logger import LOGGER
 from custom_components.hacs.validate.manager import ValidationManager
 
 TOKEN = os.getenv("INPUT_GITHUB_TOKEN")
@@ -44,11 +44,10 @@ logging.basicConfig(
     format="::%(levelname)s:: %(message)s",
     level=logging.DEBUG,
 )
-logger = get_hacs_logger()
 
 
 def error(error: str):
-    logger.error(error)
+    LOGGER.error(error)
     exit(1)
 
 
@@ -111,7 +110,7 @@ async def preflight():
         elif GITHUB_REPOSITORY == HacsGitHubRepo.DEFAULT:
             category = chose_category()
             repository = await chose_repository(hacs.githubapi, category)
-            logger.info(f"Actor: {GITHUB_ACTOR}")
+            LOGGER.info(f"Actor: {GITHUB_ACTOR}")
         else:
             category = CATEGORY.lower()
             if event_data.get("pull_request") is not None:
@@ -120,9 +119,16 @@ async def preflight():
                 repository = head["repo"]["full_name"]
             else:
                 repository = GITHUB_REPOSITORY
+                if event_data.get("ref") is not None:
+                    # For push events
+                    ref = event_data["ref"]
 
-        logger.info(f"Category: {category}")
-        logger.info(f"Repository: {repository}")
+                    # For tag events
+                    if ref.startswith("refs/tags/"):
+                        ref = ref.split("/")[-1]
+
+        LOGGER.info(f"Category: {category}")
+        LOGGER.info(f"Repository: {repository}{f'@{ref}' if ref else ''}")
 
         if TOKEN is None:
             error("No GitHub token found, use env GITHUB_TOKEN to set this.")
@@ -158,5 +164,4 @@ async def validate_repository(hacs, repository, category, ref=None):
         error(exception)
 
 
-LOOP = asyncio.get_event_loop()
-LOOP.run_until_complete(preflight())
+asyncio.run(preflight())
