@@ -48,7 +48,7 @@ class HacsPythonScriptRepository(HacsRepository):
                 break
         if not compliant:
             raise HacsException(
-                f"Repository structure for {self.ref.replace('tags/','')} is not compliant"
+                f"{self.string} Repository structure for {self.ref.replace('tags/','')} is not compliant"
             )
 
         # Handle potential errors
@@ -62,6 +62,9 @@ class HacsPythonScriptRepository(HacsRepository):
         """Registration."""
         # Set name
         self.update_filenames()
+
+        if self.hacs.system.action:
+            await self.hacs.validation.async_run_repository_checks(self)
 
     @concurrent(concurrenttasks=10, backoff_time=5)
     async def update_repository(self, ignore_issues=False, force=False):
@@ -80,7 +83,7 @@ class HacsPythonScriptRepository(HacsRepository):
                 break
         if not compliant:
             raise HacsException(
-                f"Repository structure for {self.ref.replace('tags/','')} is not compliant"
+                f"{self.string} Repository structure for {self.ref.replace('tags/','')} is not compliant"
             )
 
         # Update name
