@@ -37,6 +37,8 @@ class HacsThemeRepository(HacsRepository):
         except BaseException:  # lgtm [py/catch-base-exception] pylint: disable=broad-except
             pass
 
+        self.hacs.async_setup_frontend_endpoint_themes()
+
     async def validate_repository(self):
         """Validate."""
         # Run common validation steps.
@@ -50,7 +52,7 @@ class HacsThemeRepository(HacsRepository):
                 break
         if not compliant:
             raise HacsException(
-                f"Repository structure for {self.ref.replace('tags/','')} is not compliant"
+                f"{self.string} Repository structure for {self.ref.replace('tags/','')} is not compliant"
             )
 
         if self.repository_manifest.content_in_root:
@@ -68,6 +70,9 @@ class HacsThemeRepository(HacsRepository):
         # Set name
         self.update_filenames()
         self.content.path.local = self.localpath
+
+        if self.hacs.system.action:
+            await self.hacs.validation.async_run_repository_checks(self)
 
     @concurrent(concurrenttasks=10, backoff_time=5)
     async def update_repository(self, ignore_issues=False, force=False):
