@@ -6,11 +6,15 @@ from typing import TYPE_CHECKING
 from jinja2 import Template
 
 if TYPE_CHECKING:
+    from ..base import HacsBase
     from ..repositories.base import HacsRepository
 
 
-def render_template(content: str, context: HacsRepository) -> str:
+def render_template(hacs: HacsBase, content: str, context: HacsRepository) -> str:
     """Render templates in content."""
+    if hacs.configuration.experimental:
+        # Do not render for experimental
+        return content
     # Fix None issues
     if context.releases.last_release_object is not None:
         prerelease = context.releases.last_release_object.prerelease
@@ -27,6 +31,8 @@ def render_template(content: str, context: HacsRepository) -> str:
             version_available=context.releases.last_release,
             version_installed=context.display_installed_version,
         )
-    except BaseException as exception:  # lgtm [py/catch-base-exception] pylint: disable=broad-except
+    except (
+        BaseException  # lgtm [py/catch-base-exception] pylint: disable=broad-except
+    ) as exception:
         context.logger.debug(exception)
     return content
