@@ -20,10 +20,10 @@ class Validator(ActionValidationBase):
 
     async def async_validate(self):
         """Validate the repository."""
-        info = await self.repository.async_get_info_file_contents(self.repository.ref)
+        info = await self.repository.async_get_info_file_contents()
         for line in info.split("\n"):
             if "<img" in line or "![" in line:
                 if [ignore for ignore in IGNORED if ignore in line]:
                     continue
                 return
-        raise ValidationException("The repository does not have issues enabled")
+        raise ValidationException("The repository does not have images in the Readme file")
