@@ -13,8 +13,9 @@ class Validator(ActionValidationBase):
     """Validate the repository."""
 
     more_info = "https://hacs.xyz/docs/publish/include#check-repository"
+    allow_fork = False
 
     async def async_validate(self):
         """Validate the repository."""
         if not self.repository.data.topics:
-            raise ValidationException("The repository has no topics")
+            raise ValidationException("The repository has no valid topics")
